requirements.txt
Flask==2.3.2
openai==1.59.3
python-dotenv==1.0.0
gunicorn==20.1.0
google-cloud-storage==2.10.0